Fix to get rid of double->int cast warning.
git-svn-id: http://svn.eol.ucar.edu/svn/ka/trunk@366 3b647ff1-7a5c-455a-93f8-4e24338e3c8e

+++ src/kadrx/PulseData.cpp
@@ -96,8 +96,8 @@ void PulseData::combineEverySecondGate()
     
     // Scale firstI and firstQ to generate values for our new gate.
    
-    int newI = firstI * iqScale;
-    int newQ = firstQ * iqScale;
+    double newI = firstI * iqScale;
+    double newQ = firstQ * iqScale;
 
     if (newI < -32768) {
       newI = -32768;
